projects auth + tests

+++ src/routes/project.routes.js
@@ -1,11 +1,16 @@
 const { Router } = require('express');
 const { mongo } = require('mongoose');
 const { ProjectService } = require('../services');
+const { AuthenticationMiddleware } = require('../middlewares');
 
 const projectRouter = new Router();
 
-projectRouter.get('/', async (req, res) => {
-  const projects = await ProjectService.getAllProjects();
+projectRouter.get('/', AuthenticationMiddleware, async (req, res) => {
+  const {
+    user: { _id: userId },
+  } = req;
+
+  const projects = await ProjectService.getAllProjects(userId);
 
   res.status(200).json({
     status: 'success',
@@ -13,10 +18,15 @@ projectRouter.get('/', async (req, res) => {
   });
 });
 
-projectRouter.post('/', async (req, res) => {
+projectRouter.post('/', AuthenticationMiddleware, async (req, res) => {
+  const {
+    user: { _id: userId },
+  } = req;
+
   const { body: newProject } = req;
   if (newProject._id == null) newProject._id = mongo.ObjectId();
 
+  newProject.userId = userId;
   const project = await ProjectService.createOrUpdateProject(newProject);
 
   res.status(200).json({
@@ -25,14 +35,27 @@ projectRouter.post('/', async (req, res) => {
   });
 });
 
-projectRouter.delete('/:id', async (req, res) => {
+projectRouter.delete('/:id', AuthenticationMiddleware, async (req, res) => {
+  const {
+    user: { _id: userId },
+  } = req;
   const { id } = req.params;
-  const project = await ProjectService.deleteProject(id);
+
+  let project = await ProjectService.getOneProject(id);
+
   if (!project) {
     const err = new Error('Project not found');
     err.status = 404;
     throw err;
   }
+
+  if (project.userId != userId) {
+    const err = new Error('Unathorized');
+    err.status = 403;
+    throw err;
+  }
+
+  project = await ProjectService.deleteProject(id);
 
   res.status(200).json({
     status: 'success',
